% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{as.data.frame.ospgsa_result}
\alias{as.data.frame.ospgsa_result}
\title{Coerce GSA result indices to a data frame}
\usage{
\method{as.data.frame}{ospgsa_result}(x, ...)
}
\arguments{
\item{x}{An \code{ospgsa_result}.}

\item{...}{Unused.}
}
\value{
The tidy indices \code{data.frame}.
}
\description{
Coerce GSA result indices to a data frame
}
