% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{gsa_parameters}
\alias{gsa_parameters}
\title{Collect parameter descriptions}
\usage{
gsa_parameters(...)
}
\arguments{
\item{...}{One or more \code{\link[=gsa_parameter]{gsa_parameter()}} objects, or a single list of them.}
}
\value{
An object of class \code{gsa_parameters} (an ordered, named collection).
}
\description{
Collect parameter descriptions
}
\examples{
gsa_parameters(
  gsa_parameter("a", "uniform", min = 0, max = 1),
  gsa_parameter("b", "lognormal", median = 2, gsd = 1.5)
)
}
\seealso{
\code{\link[=gsa_parameter]{gsa_parameter()}}, \code{\link[=gsa_correlation]{gsa_correlation()}}
}
