% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ospgsa_result}
\alias{plot.ospgsa_result}
\alias{gsa_plot}
\title{Plot method for ospgsa results}
\usage{
\method{plot}{ospgsa_result}(x, type = c("indices", "sobol", "delta_two_stage", "morris", "time"), ...)

gsa_plot(x, type = "indices", ...)
}
\arguments{
\item{x}{An \link{ospgsa_result}.}

\item{type}{One of \code{"indices"}, \code{"sobol"}, \code{"delta_two_stage"}, \code{"morris"},
\code{"time"}.}

\item{...}{Passed to the specific plotting function.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot method for ospgsa results
}
